[FLINK-37701] Extract LocalRecoverTest to reduce AdaptiveSchedulerTest length, restore checkstyle line-length limit.
